Add `useStreamJSON` hook (#26)
Adds a ~~`useJSON`~~ `useStreamJSON` utility hook, to be used with large arrays of data, which normally would be too large for a command to load in memory directly.

It can be used with either http(s) URLs, or `file:///` URLs. In either case, the hook will create a stream using the URL (either using `fetch` or `createReadStream`), and then use `stream-json` to stream through it.

https://github.com/raycast/utils/assets/1155589/061e60b8-464f-45df-8a42-5172b1990377

- [x] Documentation
- [x] Add support for arrays wrapped in objects, not just arrays
- [x] Find a better name? 😅 => renamed to `useStreamJSON`